fix linux warnings

+++ Mercury2/src/MSemaphore.cpp
@@ -83,9 +83,9 @@ __sync_and_and_fetch (
     return Old & Value;
 }
 
-#define SYNC_OR_AND_FETCH(d,v) OrAndFetch(d,v)
-#define COMPARE_AND_SWAP(d,o,n) InterlockedCompareExchange(d, n, o)
-#define SYNC_AND_AND_FETCH(d,v) __sync_and_and_fetch(d,v)
+#define SYNC_OR_AND_FETCH(d,v) ((uint32_t)OrAndFetch(d,v))
+#define COMPARE_AND_SWAP(d,o,n) ((uint32_t)InterlockedCompareExchange(d, n, o))
+#define SYNC_AND_AND_FETCH(d,v) ((uint32_t)__sync_and_and_fetch(d,v))
 
 unsigned long MSemaphore::ReadAndClear()
 {
@@ -111,13 +111,13 @@ unsigned long MSemaphore::Read()
 
 void MSemaphore::WaitAndSet(uint32_t value, uint32_t newVal)
 {
-	while( COMPARE_AND_SWAP(&m_semaphore, value, newVal) != ((int32_t)value) );
+	while( COMPARE_AND_SWAP(&m_semaphore, value, newVal) != value );
 }
 
 void MSemaphore::Wait()
 {
 	uint32_t thread = MercuryThread::Current();
-	if ( SYNC_OR_AND_FETCH(&m_semaphore, 0) == ((int32_t)thread) ) //recursive lock
+	if ( SYNC_OR_AND_FETCH(&m_semaphore, 0) == thread ) //recursive lock
 	{
 		++m_lockCount;
 		return;
@@ -130,7 +130,7 @@ void MSemaphore::Wait()
 void MSemaphore::UnLock()
 {
 	uint32_t thread = MercuryThread::Current();
-	if ( SYNC_OR_AND_FETCH(&m_semaphore, 0) == ((int32_t)thread) ) //unlock given from correct thread
+	if ( SYNC_OR_AND_FETCH(&m_semaphore, 0) == thread ) //unlock given from correct thread
 	{
 		--m_lockCount;
 		if (m_lockCount == 0) SYNC_AND_AND_FETCH(&m_semaphore, 0 );
